refactor: Move SubletItem to PennMobileShared
